add sendTransaction method

+++ src/main/java/accounts/Account.java
@@ -3,7 +3,7 @@ package accounts;
 public class Account {
 
 	private Address address;    // アドレス
-	private int balance;        // 残高
+	private double balance;        // 残高
 	private int nonce;          // トランザクションを実行した回数
 
 	public Account(Address address) {
@@ -17,11 +17,11 @@ public class Account {
 	}
 	// getter
 	public Address getAddress() { return address; }
-	public int getBalance() { return balance; }
+	public double getBalance() { return balance; }
 	public int getNonce() { return nonce; }
 
 	// setter
 	public void setAddress(Address address) { this.address = address; }
-	public void setBalance(int balance) { this.balance = balance; }
+	public void setBalance(double balance) { this.balance = balance; }
 	public void setNonce(int nonce) { this.nonce = nonce; }
 }
